Split OpenSSL backend into separate files (EC)
